Somewhat functioning GameController that rotates the molecule. Not yet integrating into running program.

+++ Nucleus/core/src/com/nucleus/Model/IMolecule.java
@@ -2,8 +2,8 @@ package com.nucleus.Model;
 
 public interface IMolecule {
 
-    void setRotation(int r);
-    int getRotation();
+    void setRotation(float r);
+    float getRotation();
     boolean isFull();
     Vector rotate(Vector v1, Vector v2, double d);
